+++ src/backend/authentication/social_auth.py
@@ -1,8 +1,11 @@
-"""Accounts's custom strategies for Python Social Auth"""
+"""Accounts's custom classes for Python Social Auth."""
 
 from django.conf import settings
 
+from social_django.models import DjangoStorage
 from social_django.strategy import DjangoStrategy
+
+from authentication.models import IdentityProviderUser
 
 
 class OptionalURLSettingStrategy(DjangoStrategy):
@@ -13,3 +16,9 @@ class OptionalURLSettingStrategy(DjangoStrategy):
         if not value:  # Don't try to do stuff if the settings is falsy
             return value
         return super().get_setting(name)
+
+
+class AccountsDjangoStorage(DjangoStorage):
+    """Custom storage to encrypt the `extra_data` content."""
+
+    user = IdentityProviderUser