+++ CHANGELOG.md
@@ -14,5 +14,6 @@ and this project adheres to
 
 - ✨(users) make `email` our username
 - ✨(authentication) Support multiple identity providers
+- ✨(authentication) Encrypt identity providers' `extra_data`
 
 [unreleased]: https://github.com/suitenumerique/docs/compare/main